Sprint Management

+++ backtrackapp/views.py
@@ -25,9 +25,6 @@ def has_access(user, project_pk):
     return False
 
 
-
-
-
 #Sees if user has permission to add/modify/delete PBI
 #True if user is product owner and it is their current project
 def is_productowner(user, project_pk):
@@ -51,14 +48,6 @@ def is_dev(user, project_pk):
     return False
 
 
-#Sees if the user is the scrum master (manager)
-#True if user is scrum master
-def is_scrummaster(user):
-    if user.role == 'M':
-        return True
-    return False
-
-
 class SignUpView(generic.CreateView):
     form_class = CustomUserCreationForm
     success_url = reverse_lazy('login')
@@ -69,8 +58,9 @@ class IndexView(generic.View):
     def get(self, request):
         user = request.user
         if user.is_authenticated:
-            if is_scrummaster(user):
-                return ProjectView.get(self, request, None)
+            if user.isManager:
+                project = get_object_or_404(Project, pk=user.current_project.pk)
+                return ProjectView.get(self, request, project.pk)
             else:
                 if user.current_project != None:
                     project = get_object_or_404(Project, pk=user.current_project.pk)
@@ -89,9 +79,12 @@ class ProjectView(generic.View):
     def get(self, request, currentProjectID):
         user = request.user
         if user.is_authenticated:
-            if is_scrummaster(user):
+            if user.isManager:
                 projectTeamsManaged = ProjectTeam.objects.filter(scrum_master=user).all()
-                return render(request, 'backtrackapp/scrumMasterProject.html', {'projectTeamsManaged_list':projectTeamsManaged})
+                projectsManaged=[]
+                for pt in projectTeamsManaged:
+                    projectsManaged.append(pt.project)
+                return render(request, 'backtrackapp/scrumMasterProject.html', {'projects':projectsManaged})
             else:
                 project = get_object_or_404(Project, pk=currentProjectID)
                 if is_productowner(user, currentProjectID):
@@ -117,8 +110,8 @@ class NewProjectView(generic.View):
                 projects = Project.objects.all()
                 projectForm = NewProjectForm()
                 teamForm = ProjectTeamForm()
-                teamForm.fields['scrum_master'].queryset = User.objects.filter(current_project=None).exclude(pk=user.pk)
-                teamForm.fields['dev_team'].queryset = User.objects.filter(current_project=None).exclude(pk=user.pk)
+                teamForm.fields['scrum_master'].queryset = User.objects.filter(isManager = True).all()
+                teamForm.fields['dev_team'].queryset = User.objects.filter(current_project=None).exclude(pk=user.pk).filter(isManager = False).exclude(pk=user.pk)
 
                 return render(request, 'backtrackapp/_new_project.html', {'projectForm': projectForm, 'teamForm': teamForm})
         else:
@@ -126,6 +119,10 @@ class NewProjectView(generic.View):
 
     def post(self, request):
         user = request.user
+
+        if user.isManager:
+            return HttpResponse("Managers cannot create projects")
+
         projectForm = NewProjectForm(request.POST)
         teamForm = ProjectTeamForm(request.POST)
 
@@ -153,7 +150,7 @@ class NewProjectView(generic.View):
         dev_team.append(str(user.id))
         result = True if len(dev_team) == len(set(dev_team)) else False
 
-        for member_id in dev_team:
+        for member_id in request.POST.getlist('dev_team'):
             member = get_object_or_404(User, pk=int(member_id))
             if member.current_project != None:
                 return False
@@ -212,29 +209,30 @@ class ProductBacklogView(generic.View):
 
                 context = {'form': NewPBIForm(initial={'project': project}), 'project': project, 'latestSprint': latestSprint, 'pbi_cum_points_list': pbi_cum_points_list}
 
-                if is_scrummaster(user):
-                    return render(request, 'backtrackapp/scrumMasterProductBacklog.html', context)
-                elif is_productowner(user, project.pk):
+                if is_productowner(user, project.pk):
                     return render(request, 'backtrackapp/productOwnerProductBacklog.html', context)
-                else:
+                elif is_dev(user, project.pk):
                     return render(request, 'backtrackapp/busyDeveloperProductBacklog.html', context)
+                else:
+                    return render(request, 'backtrackapp/scrumMasterProductBacklog.html', context)
             else:
                 raise Http404("You do not have access to this project")
         else:
             return HttpResponseRedirect(reverse('login'))
 
     def post(self, request, pk):
+        print(pk)
         if request.POST.get('pbi_id') != None:
             pbi_id = int(request.POST.get('pbi_id'))
 
         if 'addToSprint' in request.POST:
             return self.addToSprint(request, pbi_id, pk)
         if 'createNewPBI' in request.POST:
-            return self.createNewPBI(request, pbi_id)
+            return self.createNewPBI(request, pbi_id, pk)
         if 'deletePBI' in request.POST:
-            return self.deletePBI(request, pbi_id)
+            return self.deletePBI(request, pbi_id, pk)
         if 'splitPBI' in request.POST:
-            return self.splitPBI(request, pbi_id)
+            return self.splitPBI(request, pbi_id, pk)
         if 'modifyPBI' in request.POST:
             return HttpResponseRedirect(reverse('backtrack:modify_PBI', args=(pbi_id,)))
         else:
@@ -249,41 +247,25 @@ class ProductBacklogView(generic.View):
         project = get_object_or_404(Project, pk=project_id)
         latestSprint = project.getLatestSprint()
 
+        if latestSprint == None:
+            return HttpResponse("You must create the sprint before adding PBIs to the current sprint")
         if latestSprint.status != 'IP':
             return HttpResponse("You must start the sprint before adding PBIs to the current sprint")
 
         pbi = get_object_or_404(ProductBacklogItem, pk=pbi_id)
         pbi.sprint = latestSprint
+        pbi.status = 'IP'
         pbi.save()
-        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(project_id,)))
+        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(project_id, )))
 
-    def deletePBI(self, request, pk):
-        pbi = get_object_or_404(ProductBacklogItem, pk=pk)
-        self.updatePrioritiesDelete(pk, pbi.priority)
+    def deletePBI(self, request, pbi_id, project_id):
+        if is_productowner(request.user, project_id) == False:
+            return HttpResponse("Only product owners can modify the product backlog")
+
+        pbi = get_object_or_404(ProductBacklogItem, pk=pbi_id)
+        self.updatePrioritiesDelete(project_id, pbi.priority)
         pbi.delete()
-        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(pk,)))
-
-    def splitPBI(self, request, pk):
-        proj = get_object_or_404(Project, pk=pk)
-        num = request.POST.get('numOfChildPBI')
-        pbi_id = request.POST.get('pbi')
-        pbi = ProductBacklogItem.objects.get(pk=pbi_id)
-        i = 1
-        while (i <= int(num)):
-            ProductBacklogItem.objects.create(name = pbi.name + "." + str(i), desc = pbi.desc, priority = pbi.priority, storypoints = pbi.storypoints, status = pbi.status, project = proj)
-            i = i + 1
-        ProductBacklogItem.objects.get(pk=pbi_id).delete()
-        #redirect back to product backlog view
-        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(pk,)))
-
-    def checkPriority(self, pk, pri):
-        proj = get_object_or_404(Project, pk=pk)
-        max_priority = proj.productbacklogitem_set.all().count() + 1
-
-        if int(pri) > max_priority or int(pri) <= 0:
-            return False
-        else:
-            return True
+        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(project_id, )))
 
     def updatePriorities(self, request, pk, pri):
         proj = get_object_or_404(Project, pk=pk)
@@ -300,18 +282,20 @@ class ProductBacklogView(generic.View):
                     i.priority = temp + 1
                     i.save()
 
-    def updatePrioritiesDelete(self, pk, pri):
-        proj = get_object_or_404(Project, pk=pk)
+    def updatePrioritiesDelete(self, proj_id, pri):
+        proj = get_object_or_404(Project, pk=proj_id)
         pbis = proj.productbacklogitem_set.all()
         for pbi in pbis:
             if pbi.priority > pri:
                 pbi.priority = pbi.priority - 1
                 pbi.save()
 
-    def splitPBI(self, request, pk):
-        proj = get_object_or_404(Project, pk=pk)
+    def splitPBI(self, request, pbi_id, project_id):
+        if is_productowner(request.user, project_id) == False:
+            return HttpResponse("Only product owners can modify the product backlog")
+
+        proj = get_object_or_404(Project, pk=project_id)
         num = request.POST.get('numOfChildPBI')
-        pbi_id = request.POST.get('pbi')
         pbi = ProductBacklogItem.objects.get(pk=pbi_id)
         i = 1
         while (i <= int(num)):
@@ -319,7 +303,7 @@ class ProductBacklogView(generic.View):
             i = i + 1
         ProductBacklogItem.objects.get(pk=pbi_id).delete()
         #redirect back to product backlog view
-        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(pk,)))
+        return HttpResponseRedirect(reverse('backtrack:project_pb', args=(project_id,)))
 
 
 
@@ -451,19 +435,21 @@ class ModifyPBIView(generic.View):
 
 
 
-
-
 #Views handling the client accessing the Sprint Backlog
 class SprintBacklogView(generic.View):
     def get(self, request, pk):
         user = request.user
+
         if user.is_authenticated:
             if has_access(user, pk):
                 project = get_object_or_404(Project, pk=pk)
                 sprint = Sprint.objects.filter(project=project).all()
                 latestSprint = project.getLatestSprint()
 
-                return self.renderSprint(request, latestSprint, project)
+                if latestSprint == None or latestSprint.status == 'C':
+                    return self.renderSprint(request, None, project)
+                else:
+                    return self.renderSprint(request, latestSprint, project)
 
             else:
                 return Http404("You do not have access to this project")
@@ -476,61 +462,84 @@ class SprintBacklogView(generic.View):
         project = get_object_or_404(Project, pk=pk)
         latestSprint = project.getLatestSprint()
 
-        #possible methods of posting
-        #getting a sprint sprint number
-
         if request.POST.get('sprint_id') != None:
             sprint_id = request.POST.get('sprint_id')
             sprint = get_object_or_404(Sprint, pk=sprint_id)
             return self.renderSprint(request, sprint, project)
+
         if request.POST.get('sprint_action') != None:
             if latestSprint == None or latestSprint.status == 'C':
-                return self.createNewSprint(project)
-            else:
-                if latestSprint.status == 'NS':
-                    project.startCurrentSprint()
-                    return HttpResponseRedirect(reverse('backtrack:project_sb', args=(latestSprint.pk, )))
-                elif latestSprint.status =='IP':
-                    project.endCurrentSprint()
-                    return HttpResponseRedirect(reverse('backtrack:project_sb', args=(latestSprint.pk, )))
+                return self.createNewSprint(request, project)
+            elif latestSprint.status == 'NS':
+                project.startCurrentSprint()
+                return HttpResponseRedirect(reverse('backtrack:project_sb', args=(project.pk, )))
+            elif latestSprint.status =='IP':
+                project.endCurrentSprint()
+                return HttpResponseRedirect(reverse('backtrack:project_sb', args=(project.pk, )))
+
 
         #what developers can do for tasks in pbi
         if request.POST.get('add_task') != None:
             pbi_id = request.POST.get('pbi_id')
             return HttpResponseRedirect(reverse('backtrack:new_task', args=(pbi_id,)))
         if request.POST.get('delete_task') != None:
-            return self.deleteTask(request, latestSprint)
+            return self.deleteTask(request, pk)
         if request.POST.get('modify_task') != None:
             return self.modifyTask(request, pk)
         if request.POST.get('remove_pbi') != None:
-            return self.removePBI(request, latestSprint)
+            return self.removePBI(request, pk)
+        if request.POST.get('complete_pbi') != None:
+            return self.completePBI(request, pk)
+        if request.POST.get('complete_task') != None:
+            return self.completeTask(request, pk)
 
-    def deleteTask(self, request, latestSprint):
+    def deleteTask(self, request, pk):
         task_id = request.POST.get('task_id')
         Task.objects.get(pk=task_id).delete()
-        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(latestSprint.pk, )))
+        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(pk, )))
 
-    def removePBI(self, request, latestSprint):
+    def completeTask(self, request, pk):
+        task_id = request.POST.get('task_id')
+        task = get_object_or_404(Task, pk=task_id)
+        task.completeTask()
+        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(pk, )))
+
+    def removePBI(self, request, pk):
         pbi_id = request.POST.get('pbi_id')
         pbi = get_object_or_404(ProductBacklogItem, pk=pbi_id)
         pbi.sprint = None;
         pbi.save()
-        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(latestSprint.pk, )))
+        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(pk, )))
+
+    def completePBI(self, request, pk):
+        pbi_id = request.POST.get('pbi_id')
+        pbi = get_object_or_404(ProductBacklogItem, pk=pbi_id)
+        pbi.status = 'C'
+        pbi.save()
+        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(pk, )))
 
     def modifyTask(self, request, pk):
         task_id = request.POST.get('task_id')
         return HttpResponseRedirect(reverse('backtrack:modify_task', args=(task_id,)))
 
-    def createNewSprint(self, project):
+    def createNewSprint(self, request, project):
+        capacity = request.POST.get('capacity')
+        print(capacity)
         newSprint = project.createNewSprint()
+        newSprint.totalCapacity = capacity
         newSprint.save()
-        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(newSprint.pk,)))
+        return HttpResponseRedirect(reverse('backtrack:project_sb', args=(project.pk,)))
 
     def renderSprint(self, request, sprint, project):
+        if sprint == None:
+            context = {'project': project, 'status':1}
+            if is_dev(request.user, project.pk):
+                return render(request, 'backtrackapp/developerNoneSprintBacklog.html', context)
+            else:
+                return render(request, 'backtrackapp/noneSprintBacklog.html', context)
+
         user = request.user
-
         tasksInAllSprintPBIs = []
-
         sprintPBI_set = ProductBacklogItem.objects.filter(sprint=sprint).all()
 
         for sprintPBI in sprintPBI_set:
@@ -539,6 +548,7 @@ class SprintBacklogView(generic.View):
                 tasksInAllSprintPBIs.append(task)
 
         latestSprint = project.getLatestSprint()
+
         if latestSprint == None or latestSprint.status == 'C':
             status = 1
         elif latestSprint.status == 'NS':
@@ -548,12 +558,10 @@ class SprintBacklogView(generic.View):
 
         context = {'sprint':sprint, 'project': project, 'sprintPBI_set': sprintPBI_set, 'status':status}
 
-        if is_scrummaster(user):
-            return render(request, 'backtrackapp/scrumMasterSprintBacklog.html', context)
-        elif is_productowner(user, project.pk):
-            return render(request, 'backtrackapp/productOwnerSprintBacklog.html', context)
+        if is_dev(user, project.pk) and sprint == latestSprint:
+            return render(request, 'backtrackapp/modificationSprintBacklog.html', context)
         else:
-            return render(request, 'backtrackapp/busyDeveloperSprintBacklog.html', context)
+            return render(request, 'backtrackapp/nomodificationSprintBacklog.html', context)
